+++ src/main/java/speiger/src/collections/bytes/maps/interfaces/Byte2DoubleMap.java
@@ -94,7 +94,7 @@ public interface Byte2DoubleMap extends Map<Byte, Double>, Byte2DoubleFunction
 	 * @return the last present value or default return	value.
 	 */
 	public default Double put(Map.Entry<Byte, Double> entry) {
-		return put(entry.getKey(), entry.getValue());
+		return put(entry.getKey().byteValue(), entry.getValue().doubleValue());
 	}
 
 	/**
