+++ src/main/java/speiger/src/collections/doubles/maps/interfaces/Double2IntMap.java
@@ -94,7 +94,7 @@ public interface Double2IntMap extends Map<Double, Integer>, Double2IntFunction
 	 * @return the last present value or default return	value.
 	 */
 	public default Integer put(Map.Entry<Double, Integer> entry) {
-		return put(entry.getKey(), entry.getValue());
+		return put(entry.getKey().doubleValue(), entry.getValue().intValue());
 	}
 
 	/**
