+++ src/main/java/speiger/src/collections/longs/maps/interfaces/Long2ObjectMap.java
@@ -95,7 +95,7 @@ public interface Long2ObjectMap<V> extends Map<Long, V>, LongFunction<V>
 	 * @return the last present value or default return	value.
 	 */
 	public default V put(Map.Entry<Long, V> entry) {
-		return put(entry.getKey(), entry.getValue());
+		return put(entry.getKey().longValue(), entry.getValue());
 	}
 
 	/**
